redo all the integration tests
