--- integration/airflow/tests/failures/server/Dockerfile
@@ -1,6 +1,0 @@
-FROM python:3.9-slim
-WORKDIR /app
-COPY requirements.txt requirements.txt
-RUN pip install -r requirements.txt
-COPY . .
-CMD ["flask", "run", "--host=0.0.0.0"]
