--- integration/airflow/tests/failures/server/requirements.txt
@@ -1,1 +1,0 @@
-flask==2.0.1
