--- integration/airflow/tests/integration/docker/entrypoint.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-#
-# SPDX-License-Identifier: Apache-2.0.
-#
-# Usage: $ ./entrypoint.sh
-
-set -e
-
-python test_integration.py
