+++ integration/airflow/tests/integration/tests/docker-compose-2.yml
@@ -2,67 +2,59 @@ version: "3.7"
 
 x-airflow-base: &airflow-base
   build:
-    context: .
+    context: ../
     target: airflow
-  environment:
+    dockerfile: Dockerfile.2
+    args:
+      AIRFLOW_IMAGE: ${AIRFLOW_IMAGE}
+  environment: &airflow-common-env
     DB_BACKEND: postgresql+psycopg2
     DB_HOST: postgres
     DB_PORT: 5432
-    AIRFLOW_FERNET_KEY: Z2uDm0ZL60fXNkEXG8LW99Ki2zf8wkmIltaTz1iQPDU=
     AIRFLOW_DATABASE_HOST: postgres
     AIRFLOW_DATABASE_NAME: airflow
     AIRFLOW_DATABASE_USERNAME: airflow
     AIRFLOW_DATABASE_PASSWORD: airflow
-    AIRFLOW_EXECUTOR: CeleryExecutor
-    AIRFLOW_LOAD_EXAMPLES: "no"
+    AIRFLOW_FERNET_KEY: Z2uDm0ZL60fXNkEXG8LW99Ki2zf8wkmIltaTz1iQPDU=
+    AIRFLOW__CORE__SQL_ALCHEMY_CONN: postgresql+psycopg2://airflow:airflow@postgres:5432/airflow
+    AIRFLOW__CELERY__RESULT_BACKEND: db+postgresql://airflow:airflow@postgres/airflow
+    AIRFLOW__CELERY__BROKER_URL: redis://:@redis:6379/0
+    AIRFLOW__CORE__EXECUTOR: CeleryExecutor
+    AIRFLOW__CORE__LOAD_EXAMPLES: "False"
+    AIRFLOW__CORE__DAGBAG_IMPORT_TIMEOUT: 120
     AIRFLOW__LINEAGE__BACKEND: openlineage.lineage_backend.OpenLineageBackend
     AIRFLOW_CONN_FOOD_DELIVERY_DB: postgres://food_delivery:food_delivery@postgres:5432/food_delivery
     AIRFLOW_CONN_BQ_CONN: google-cloud-platform://?extra__google_cloud_platform__project=openlineage-ci&extra__google_cloud_platform__key_path=%2Fopt%2Fconfig%2Fgcloud%2Fgcloud-service-key.json
-    AIRFLOW__CELERY__BROKER_URL: redis://redis:6379/0
-    AIRFLOW__CORE__LOGGING_CONFIG_CLASS: log_config.LOGGING_CONFIG
-    AIRFLOW__CORE__DAGS_ARE_PAUSED_AT_CREATION: "false"
-    AIRFLOW__CORE__SQL_ALCHEMY_CONN: postgresql+psycopg2://airflow:airflow@postgres:5432
+    AIRFLOW__CORE__DAGS_ARE_PAUSED_AT_CREATION: "False"
+    BIGQUERY_PREFIX: ${BIGQUERY_PREFIX}
+    BIGQUERY_DBT_DATASET: ${BIGQUERY_DBT_DATASET}_dbt
     GOOGLE_APPLICATION_CREDENTIALS: /opt/config/gcloud/gcloud-service-key.json
-    OPENLINEAGE_NAMESPACE: food_delivery
     OPENLINEAGE_URL: http://backend:5000
+    OPENLINEAGE_NAMESPACE: food_delivery
     OPENLINEAGE_EXTRACTOR_CustomOperator: custom_extractor.CustomExtractor
-    BIGQUERY_PREFIX: airflow_1
-    BIGQUERY_DBT_DATASET: airflow_1_dbt
   volumes:
-    - ./airflow/config/log_config.py:/opt/airflow/config/log_config.py
-    - ./airflow/dags:/opt/airflow/dags
-    - ./airflow/logs:/opt/airflow/logs
-    - ./gcloud:/opt/config/gcloud
+      - ./airflow/config/log_config.py:/opt/airflow/config/log_config.py
+      - ./airflow/logs:/opt/airflow/logs
+      - ./airflow/dags:/opt/airflow/dags
+      - ../gcloud:/opt/config/gcloud
 
 services:
   integration:
     build:
-      context: .
+      context: ../
       target: integration
+      dockerfile: Dockerfile.2
+      args:
+        AIRFLOW_IMAGE: ${AIRFLOW_IMAGE}
+    environment:
+      BIGQUERY_PREFIX: ${BIGQUERY_PREFIX}
+      BIGQUERY_DBT_DATASET: ${BIGQUERY_DBT_DATASET}_dbt
     volumes:
       - ./docker/wait-for-it.sh:/wait-for-it.sh
     depends_on:
-      - airflow
-      - backend
-    environment:
-      - BIGQUERY_PREFIX=airflow_1
-      - BIGQUERY_DBT_DATASET=airflow_1_dbt
-    entrypoint: ["./wait-for-it.sh", "postgres:5432", "--", "./entrypoint.sh"]
-
-  airflow:
-    <<: *airflow-base
-    ports:
-      - "8080:8080"
-    command: webserver
-    depends_on:
       - airflow_scheduler
-      - postgres
-    healthcheck:
-      test: [ "CMD", "curl", "--fail", "http://localhost:8080/health" ]
-      interval: 10s
-      timeout: 10s
-      retries: 5
-    restart: always
+      - backend
+    entrypoint: ["./wait-for-it.sh", "backend:5000", "--", "python", "test_integration.py"]
 
   airflow_scheduler:
     <<: *airflow-base
@@ -76,24 +68,35 @@ services:
 
   airflow_worker:
     <<: *airflow-base
-    command: worker
+    command: celery worker
     restart: always
     healthcheck:
       test: [ "CMD-SHELL", 'celery --app airflow.executors.celery_executor.app inspect ping -d "celery@$${HOSTNAME}"' ]
       interval: 10s
       timeout: 10s
       retries: 5
+
   backend:
     build:
-      context: server
-      dockerfile: Dockerfile
+      context: ../server
     environment:
       - FLASK_APP=app.py
       - FLASK_RUN_PORT=5000
 
   airflow_init:
     <<: *airflow-base
-    command: initdb
+    command: db init
+
+  redis:
+    image: redis:latest
+    ports:
+      - 6379:6379
+    healthcheck:
+      test: ["CMD", "redis-cli", "ping"]
+      interval: 5s
+      timeout: 30s
+      retries: 50
+    restart: always
 
   postgres:
     image: bitnami/postgresql:12.1.0
@@ -110,15 +113,4 @@ services:
       - FOOD_DELIVERY_DB=food_delivery
       - ALLOW_EMPTY_PASSWORD=yes
     volumes:
-      - ./docker/init-db.sh:/docker-entrypoint-initdb.d/init-db.sh
-
-  redis:
-    image: redis:latest
-    ports:
-      - 6379:6379
-    healthcheck:
-      test: ["CMD", "redis-cli", "ping"]
-      interval: 5s
-      timeout: 30s
-      retries: 50
-    restart: always
+      - ../docker/init-db.sh:/docker-entrypoint-initdb.d/init-db.sh
